fix local variable references.
these are always (right?) 32-bit offsets.  we could change our
interface, but maybe some other architecture needs 64-bit
offsets... actually, that really doesn't make sense.  whatever.

just cast to 32-bit first, so that it properly detects negative
offsets.

+++ command.go
@@ -794,7 +794,7 @@ func symexec(inferior *ptrace.Tracee, addr uintptr) ([]register_file, error) {
           rf[movregtarget(ixn)] = memaddr(int64(addr) + int64(ixn.Len) +
                                           memref.Disp)
         } else if memref.Base == x86asm.RBP {
-          rf[movregtarget(ixn)] = lvar(memref.Disp)
+          rf[movregtarget(ixn)] = lvar(int32(memref.Disp))
         } else {
           return nil, fmt.Errorf("unknown rel mem reference in %v", ixn)
         }
